# biometric_api/docs/openapi-schema.yaml
openapi: 3.0.3
info:
  title: Biometric API
  version: 1.0.0
  description: API para administración de equipos biomédicos
paths:
...
  /api/v1/equipment/:
    get:
      operationId: v1_equipment_list
      description: |-
        CRUD de equipos biomédicos.

        Incluye:

        - CRUD
        - búsqueda por asset_tag
        - filtros
        - ordenamiento
        - regeneración de QR
        - historial de mantenimientos
      parameters:
      - in: query
        name: branch
        schema:
          type: number
      - in: query
        name: brand
        schema:
          type: number
      - in: query
        name: equipment_model
        schema:
          type: number
      - name: ordering
        required: false
        in: query
        description: Qué campo usar para ordenar los resultados.
        schema:
          type: string
      - name: page
        required: false
        in: query
        description: Un número de página dentro del conjunto de resultados paginado.
        schema:
          type: integer
      - in: query
        name: purchase_date_after
        schema:
          type: string
          format: date
      - in: query
        name: purchase_date_before
        schema:
          type: string
          format: date
      - in: query
        name: risk_class
        schema:
          type: string
      - in: query
        name: risk_class__isnull
        schema:
          type: boolean
      - name: search
        required: false
        in: query
        description: Un término de búsqueda.
        schema:
          type: string
      - in: query
        name: status
        schema:
          type: string
      tags:
      - v1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedEquipmentList'
          description: ''
    post:
      operationId: v1_equipment_create
      description: |-
        CRUD de equipos biomédicos.

        Incluye:

        - CRUD
        - búsqueda por asset_tag
        - filtros
        - ordenamiento
        - regeneración de QR
        - historial de mantenimientos
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
          description: ''
...
  /api/v1/equipment/{id}/history/:
    get:
      operationId: v1_equipment_history_retrieve
      description: Historial paginado de mantenimientos del equipo.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este Equipo biomédico.
        required: true
      tags:
      - v1
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
          description: ''
  /api/v1/equipment/{id}/regenerate-qr/:
    post:
      operationId: v1_equipment_regenerate_qr_create
      description: Regenera el código QR del equipo.
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: Un valor de entero único que identifique este Equipo biomédico.
        required: true
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
          description: ''
...
  /api/v1/equipment/regenerate-qr-all/:
    post:
      operationId: v1_equipment_regenerate_qr_all_create
      description: |-
        Regenera el código QR de todos los equipos.

        Útil tras cambiar FRONTEND_BASE_URL o al preparar una tanda de
        etiquetas para imprimir. Con ``?missing=1`` solo cubre los equipos
        que aún no tienen QR.
      tags:
      - v1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/EquipmentRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Equipment'
          description: ''
...
components:
  schemas:
...
    BlankEnum:
      enum:
      - ''
...
    Equipment:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          title: Nombre
          maxLength: 150
        asset_tag:
          type: string
          title: Placa / código de inventario
          maxLength: 50
        internal_code:
          type: string
          title: Código N.T.
          maxLength: 50
        serial:
          type: string
          title: Serie
          maxLength: 80
        software_identifier:
          type: string
          title: Identificador Software (Id)
          maxLength: 50
        equipment_model:
          type: integer
          title: Modelo
        equipment_model_name:
          type: string
          readOnly: true
        brand_name:
          type: string
          readOnly: true
        branch:
          type: integer
          title: Sede
        branch_name:
          type: string
          readOnly: true
        location:
          type: string
          title: Ubicación
          maxLength: 120
        technology_type:
          nullable: true
          title: Tipo tecnología
          oneOf:
          - $ref: '#/components/schemas/TechnologyTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        biomedical_classification:
          type: string
          title: Clasificación biomédica
          maxLength: 50
        risk_class:
          nullable: true
          title: Clasificación de riesgo INVIMA
          oneOf:
          - $ref: '#/components/schemas/RiskClassEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        risk_class_display:
          type: string
          readOnly: true
        manufacturer:
          type: string
          title: Fabricante
          maxLength: 150
        owner:
          type: string
          title: Propietario
          maxLength: 150
        client_name:
          type: string
          title: Cliente
          maxLength: 180
        branch_text:
          type: string
          title: Marca (texto)
          description: Útil si la marca no existe en catálogo.
          maxLength: 120
        department:
          type: string
          title: Departamento
          maxLength: 80
        city:
          type: string
          title: Ciudad
          maxLength: 80
        area:
          type: string
          title: Área
          maxLength: 120
        purchase_date:
          type: string
          format: date
          nullable: true
          title: Fecha de compra
        supplier_acquisition:
          type: string
          title: Proveedor de adquisión
          maxLength: 150
        equipment_cost:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          nullable: true
          title: Costo equipo
        manufacture_date:
          type: string
          format: date
          nullable: true
          title: Fecha de fabricación
        start_use_date:
          type: string
          format: date
          nullable: true
          title: Fecha inicia funcionamiento
        warranty_start_date:
          type: string
          format: date
          nullable: true
          title: Fecha inicia garantía
        warranty_end_date:
          type: string
          format: date
          nullable: true
          title: Fecha finaliza garantía
        maintenance_provider:
          type: string
          title: Proveedor de mantenimiento
          maxLength: 150
        maintenance_frequency_months:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
          title: Frecuencia mantenimiento (meses)
        last_preventive:
          type: string
          title: Último preventivo
          maxLength: 120
        next_preventive:
          type: string
          title: Próximo preventivo
          maxLength: 120
        calibration_date:
          type: string
          title: Código de calibración
          maxLength: 50
        calibration_frequency_months:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
          title: Frecuencia calibración (meses)
        last_calibration:
          type: string
          title: Última calibración
          maxLength: 120
        next_calibration:
          type: string
          title: Próxima calibración
          maxLength: 120
        corrective_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          title: Número de correctivos.
        electrical_safety_class:
          type: string
          title: Clase seguridad eléctrica
          maxLength: 80
        electrical_safety_type:
          type: string
          title: Tipo seguridad eléctrica
          maxLength: 80
        invima_registration:
          type: string
          title: Registro Invima
          maxLength: 80
        ecri:
          type: string
          maxLength: 80
        life_use_years:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
          title: Vida útil (años)
        observations:
          type: string
          title: Observaciones
        equipment_image:
          type: string
          format: uri
          nullable: true
          title: Imagen del equipo
        life_sheet_pdf:
          type: string
          format: uri
          nullable: true
          title: Hoja Vida PDF
        status:
          allOf:
          - $ref: '#/components/schemas/EquipmentStatusEnum'
          title: Estado
        status_display:
          type: string
          readOnly: true
        mtbf_hours:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          title: MTBF (horas)
          description: Tiempo promedio entre fallas consecutivas. Recalculado automáticamente.
        mttr_hours:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          readOnly: true
          nullable: true
          title: MTTR (horas)
          description: Tiempo promedio para resolver una falla. Recalculado automáticamente.
        qr_code:
          type: string
          format: uri
          nullable: true
          readOnly: true
          title: Código QR
        qr_code_url:
          type: string
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
          title: Creado
        updated_at:
          type: string
          format: date-time
          readOnly: true
          title: Actualizado
      required:
      - asset_tag
      - branch
      - branch_name
      - brand_name
      - created_at
      - equipment_model
      - equipment_model_name
      - id
      - mtbf_hours
      - mttr_hours
      - name
      - qr_code
      - qr_code_url
      - risk_class_display
      - status_display
      - updated_at
...
    EquipmentRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          title: Nombre
          maxLength: 150
        asset_tag:
          type: string
          minLength: 1
          title: Placa / código de inventario
          maxLength: 50
        internal_code:
          type: string
          title: Código N.T.
          maxLength: 50
        serial:
          type: string
          title: Serie
          maxLength: 80
        software_identifier:
          type: string
          title: Identificador Software (Id)
          maxLength: 50
        equipment_model:
          type: integer
          title: Modelo
        branch:
          type: integer
          title: Sede
        location:
          type: string
          title: Ubicación
          maxLength: 120
        technology_type:
          nullable: true
          title: Tipo tecnología
          oneOf:
          - $ref: '#/components/schemas/TechnologyTypeEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        biomedical_classification:
          type: string
          title: Clasificación biomédica
          maxLength: 50
        risk_class:
          nullable: true
          title: Clasificación de riesgo INVIMA
          oneOf:
          - $ref: '#/components/schemas/RiskClassEnum'
          - $ref: '#/components/schemas/BlankEnum'
          - $ref: '#/components/schemas/NullEnum'
        manufacturer:
          type: string
          title: Fabricante
          maxLength: 150
        owner:
          type: string
          title: Propietario
          maxLength: 150
        client_name:
          type: string
          title: Cliente
          maxLength: 180
        branch_text:
          type: string
          title: Marca (texto)
          description: Útil si la marca no existe en catálogo.
          maxLength: 120
        department:
          type: string
          title: Departamento
          maxLength: 80
        city:
          type: string
          title: Ciudad
          maxLength: 80
        area:
          type: string
          title: Área
          maxLength: 120
        purchase_date:
          type: string
          format: date
          nullable: true
          title: Fecha de compra
        supplier_acquisition:
          type: string
          title: Proveedor de adquisión
          maxLength: 150
        equipment_cost:
          type: string
          format: decimal
          pattern: ^-?\d{0,12}(?:\.\d{0,2})?$
          nullable: true
          title: Costo equipo
        manufacture_date:
          type: string
          format: date
          nullable: true
          title: Fecha de fabricación
        start_use_date:
          type: string
          format: date
          nullable: true
          title: Fecha inicia funcionamiento
        warranty_start_date:
          type: string
          format: date
          nullable: true
          title: Fecha inicia garantía
        warranty_end_date:
          type: string
          format: date
          nullable: true
          title: Fecha finaliza garantía
        maintenance_provider:
          type: string
          title: Proveedor de mantenimiento
          maxLength: 150
        maintenance_frequency_months:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
          title: Frecuencia mantenimiento (meses)
        last_preventive:
          type: string
          title: Último preventivo
          maxLength: 120
        next_preventive:
          type: string
          title: Próximo preventivo
          maxLength: 120
        calibration_date:
          type: string
          title: Código de calibración
          maxLength: 50
        calibration_frequency_months:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
          title: Frecuencia calibración (meses)
        last_calibration:
          type: string
          title: Última calibración
          maxLength: 120
        next_calibration:
          type: string
          title: Próxima calibración
          maxLength: 120
        corrective_count:
          type: integer
          maximum: 2147483647
          minimum: 0
          title: Número de correctivos.
        electrical_safety_class:
          type: string
          title: Clase seguridad eléctrica
          maxLength: 80
        electrical_safety_type:
          type: string
          title: Tipo seguridad eléctrica
          maxLength: 80
        invima_registration:
          type: string
          title: Registro Invima
          maxLength: 80
        ecri:
          type: string
          maxLength: 80
        life_use_years:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
          title: Vida útil (años)
        observations:
          type: string
          title: Observaciones
        equipment_image:
          type: string
          format: binary
          nullable: true
          title: Imagen del equipo
        life_sheet_pdf:
          type: string
          format: binary
          title: Hoja Vida PDF
        status:
          allOf:
          - $ref: '#/components/schemas/EquipmentStatusEnum'
          title: Estado
      required:
      - asset_tag
      - branch
      - equipment_model
      - name
    EquipmentStatusEnum:
      enum:
      - ACTIVE
      - INACTIVE
      - IN_MAINTENANCE
      - IN_REPAIR
      type: string
      description: |-
        * `ACTIVE` - Operativo
        * `INACTIVE` - Fuera de servicio
        * `IN_MAINTENANCE` - En mantenimiento
        * `IN_REPAIR` - En reparación
...
    NullEnum:
      enum:
      - null
...
    PaginatedEquipmentList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Equipment'
...
    RiskClassEnum:
      enum:
      - I
      - IIA
      - IIB
      - III
      type: string
      description: |-
        * `I` - Clase I — riesgo bajo
        * `IIA` - Clase IIA — riesgo moderado
        * `IIB` - Clase IIB — riesgo moderado-alto
        * `III` - Clase III — riesgo alto
...
    TechnologyTypeEnum:
      enum:
      - ELECTRONIC
      - ELECTROMEDICAL
      - MECHANICAL
      - MIXED
      - OTHER
      type: string
      description: |-
        * `ELECTRONIC` - Electrónico
        * `ELECTROMEDICAL` - Electromédico
        * `MECHANICAL` - Mecánico
        * `MIXED` - Mixto
        * `OTHER` - Otro
...